% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbl2int.R
\name{dbl2int}
\alias{dbl2int}
\title{Convert double to int}
\usage{
dbl2int(x)
}
\arguments{
\item{x}{A single double value.}
}
\value{
\code{x}, converted to integer.
}
\description{
Convert double to int
}
\examples{
dbl2int(1.2)
}
