% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testbigstatsr-package.R
\docType{package}
\name{testbigstatsr-package}
\alias{testbigstatsr}
\alias{testbigstatsr-package}
\title{testbigstatsr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
